Extract sourcesVersionedJson.ts
